requirements.txt
beautifulsoup4==4.10.0
certifi==2025.1.31
charset-normalizer==2.0.12
idna==3.10
mysql-connector-python==8.0.27
protobuf==6.30.0
requests==2.26.0
soupsieve==2.6
urllib3==1.26.20
